wrote code to determine who gets a discarded tile
-implemented Turn::determineOrderOfClaims to determine who, if anyone
gets the discard
-fixed bugs for a working game
a. made sure human player's providedInput member was toggled when
HumanPlayer::hasProvidedInput was called and returned true
b. made sure that if HumanPlayer or AIPlayer overrid a method of Player
class, that the declaration was exactly the same (HumanPlayer::makeMeld
was having this problem) and that the function was virtual in the Player
class declaration
-made sure to clear the map HandEvaluator::handClassifications whenever
splitTiles was called, otherwise all the past indices were stored even
though the hand had changed
-HandEvaluator::canDeclareBigMeldedKang and
HandEvaluator::canDeclareMeldedPeng were automatically returning a
vector of size one, regardless of whether the one meld it contained was
initialized
-updated meld_strings array of strings defined in HandEvaluator.h to
correspond to every enum of type MeldType, to avoid an out of range
error
-updated HandEvaluator::canDeclareMeledChi to account for the fact that
although the algorithm looks at two tiles in the hand at the same time,
in the case that the discarded tile is the same as the higher suit, to
account for the possibility that the index i+2 is outside of the vector
indices or an index outside of the range of the hand
-made sure to initialized TIle *Game::curr_discard to a nullptr
-made sure to clear the map of updates in the Draw class so that past
messages to the user do not show up